requirements.txt
altair==5.5.0
auth0_python==3.24.1
folium==0.19.5
matplotlib==3.5.2
numpy==1.23.3
pandas==1.5.0
plotly==5.17.0
pycountry==24.6.1
pymongo==4.11.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.1.2
seaborn==0.13.2
streamlit==1.43.2
streamlit_folium==0.24.0
streamlit_plotly_events==0.0.6
wordcloud==1.9.4
google-genai
boto3